chore: refactored buildJob into separate module
