+++ src/index.js
@@ -10,7 +10,6 @@ const isSANB = require('is-string-and-not-blank');
 const isValidPath = require('is-valid-path');
 const later = require('@breejs/later');
 const threads = require('bthreads');
-const { boolean } = require('boolean');
 const { setTimeout, setInterval } = require('safe-timers');
 
 const {
@@ -19,6 +18,7 @@ const {
   getHumanToMs,
   parseValue
 } = require('./job-utils');
+const buildJob = require('./job-builder');
 
 // bthreads requires us to do this for web workers (see bthreads docs for insight)
 threads.Buffer = Buffer;
@@ -194,119 +194,6 @@ class Bree extends EventEmitter {
   }
 
   // eslint-disable-next-line complexity
-  buildJob(job, config) {
-    if (isSANB(job)) {
-      const path = join(
-        config.root,
-        job.endsWith('.js') || job.endsWith('.mjs')
-          ? job
-          : `${job}.${config.defaultExtension}`
-      );
-
-      return {
-        name: job,
-        path,
-        timeout: config.timeout,
-        interval: config.interval
-      };
-    }
-
-    if (typeof job === 'function') {
-      const path = `(${job.toString()})()`;
-
-      return {
-        name: job.name,
-        path,
-        worker: { eval: true },
-        timeout: config.timeout,
-        interval: config.interval
-      };
-    }
-
-    // process job.path
-    if (typeof job.path === 'function') {
-      const path = `(${job.path.toString()})()`;
-
-      job.path = path;
-      job.worker = {
-        eval: true,
-        ...job.worker
-      };
-    } else {
-      const path = isSANB(job.path)
-        ? job.path
-        : join(
-            config.root,
-            job.name.endsWith('.js') || job.name.endsWith('.mjs')
-              ? job.name
-              : `${job.name}.${config.defaultExtension}`
-          );
-
-      if (isValidPath(path)) {
-        job.path = path;
-      } else {
-        // assume that it's a transformed eval string
-        job.worker = {
-          eval: true,
-          ...job.worker
-        };
-      }
-    }
-
-    if (typeof job.timeout !== 'undefined') {
-      job.timeout = this.parseValue(job.timeout);
-    }
-
-    if (typeof job.interval !== 'undefined') {
-      job.interval = this.parseValue(job.interval);
-    }
-
-    // build cron
-    if (typeof job.cron !== 'undefined') {
-      if (this.isSchedule(job.cron)) {
-        job.interval = job.cron;
-        // delete job.cron;
-      } else {
-        job.interval = later.parse.cron(
-          job.cron,
-          boolean(
-            typeof job.hasSeconds === 'undefined'
-              ? config.hasSeconds
-              : job.hasSeconds
-          )
-        );
-      }
-    }
-
-    // if timeout was undefined, cron was undefined,
-    // and date was undefined then set the default
-    // (as long as the default timeout is >= 0)
-    if (
-      Number.isFinite(config.timeout) &&
-      config.timeout >= 0 &&
-      typeof job.timeout === 'undefined' &&
-      typeof job.cron === 'undefined' &&
-      typeof job.date === 'undefined' &&
-      typeof job.interval === 'undefined'
-    )
-      job.timeout = config.timeout;
-
-    // if interval was undefined, cron was undefined,
-    // and date was undefined then set the default
-    // (as long as the default interval is > 0, or it was a schedule, or it was valid)
-    if (
-      ((Number.isFinite(config.interval) && config.interval > 0) ||
-        isSchedule(config.interval)) &&
-      typeof job.interval === 'undefined' &&
-      typeof job.cron === 'undefined' &&
-      typeof job.date === 'undefined'
-    )
-      job.interval = config.interval;
-
-    return job;
-  }
-
-  // eslint-disable-next-line complexity
   validateJob(job, i, isAdd = false) {
     const errors = [];
     const names = [];
@@ -368,7 +255,7 @@ class Bree extends EventEmitter {
           throw new Error(`Job #${i + 1} "${job}" path missing: ${path}`);
       }
 
-      return this.buildJob(job, this.config);
+      return buildJob(job, this.config);
     }
 
     // job is a function
@@ -382,7 +269,7 @@ class Bree extends EventEmitter {
 
       if (errors.length > 0) throw combineErrors(errors);
 
-      return this.buildJob(job, this.config);
+      return buildJob(job, this.config);
     }
 
     // use a prefix for errors
@@ -591,7 +478,7 @@ class Bree extends EventEmitter {
 
     if (errors.length > 0) throw combineErrors(errors);
 
-    return this.buildJob(job, this.config);
+    return buildJob(job, this.config);
   }
 
   getWorkerMetadata(name, meta = {}) {
